moved useQuery logic into custom hooks
